backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scipy
strands-agents[mcp]
strands-agents-tools
boto3
